solutions for 'remove_duplicates_from_sorted_array'

+++ src/single_number/solution2.cpp
@@ -8,7 +8,7 @@ class Solution {
 public:
     int singleNumber(int A[], int n) {
         int x = A[0];
-        for (int i=1; i != n; ++i) {
+        for (size_t i=1; i != n; ++i) {
             x ^= A[i];
         }
         return x;
